Moving hyperlink files into hyperlinks directory.
